first step of rework
